requirements.txt
streamlit
streamlit_option_menu
pandas
psycopg2
hashlib
datetime
plotly.express